ignore: add switch to read gitignores outside git repos
As mentioned in #1414, ripgrep by default does not use gitignore
patterns when outside a git repository. This is a good default, as it
prevents filtering files from non-repositories (like /tmp).

However, there are several use cases for respecting gitignore patterns
outside of a proper git repository. For example, if a git repository is
copied without the .git directory (i.e. downloaded from GitHub as a zip
file), or in SCM systems such as Perforce where ignore patterns can be
used (and .gitignore can be the name of the ignore files).

This commit introduces a "--ignore-outside-git" flag which enables
parsing gitignore files outside of proper git repositories.