ripgrep: expose custom ignore files via --ignore-file-name
The ignore crate has the concept of "custom ignore" files. These files
take precedence over all others and can be used to declare arbitrary
file names as files which should be parsed for ignore patterns.

This commit adds the --ignore-file-name switch to ripgrep, exposing this
functionality to the user. For instance, if a user wanted to put their
ignores in many files called ".my_ignore" they could:

    rg --ignore-file-name .my_ignore

This can also be used as a hack to honor .gitignore files outside .git
repositories, which can be useful if a git repo has been copied
somewhere without the .git directory (i.e. downloaded as a zip from
GitHub) or for SCM systems like Perforce which allow .gitignore files
but are not valid git repositories. Note, however, that this tactic
gives .gitignore precedence over .ignore and .rgignore, so consider
using multiple flags:

    rg --ignore-file-name .ignore --ignore-file-name .gitignore